Dev

See merge request s09-webmobile1-sub2/S09P12B303!37

+++ Server/src/main/java/com/threeracha/gaewoonhae/db/repository/EmojiRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface EmojiRepository extends JpaRepository<Emoji, Integer> {
+public interface EmojiRepository extends JpaRepository<Emoji, Long> {
     Optional<Emoji> findByEmojiIdAndEmojiPrice(int emojiId, int emojiPrice);
 
 }
